Fix remaining imports for moving rest user settings hbase client.
